requirements.txt
celery[redis]==5.6.3
coverage==7.13.5
cfgv==3.5.0
cryptography==47.0.0
cssbeautifier==1.15.4
dj-database-url==3.1.2
dj-rest-auth==7.2.0
django-allauth==65.16.1
django-cors-headers==4.9.0
django-import-export==4.4.0
django-stubs-ext==6.0.3
django-stubs==6.0.3
django-unfold==0.91.0
Django==6.0.4
djangorestframework-simplejwt==5.5.1
djangorestframework==3.17.1
djlint==1.36.4
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
drf-standardized-errors==0.15.0
editorconfig==0.17.1
feedparser==6.0.12
gunicorn==25.3.0
httpx==0.28.1
identify==2.6.19
inflection==0.5.1
jsbeautifier==1.15.4
langchain-core==1.3.2
langchain-protocol==0.0.12
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.12
langgraph-sdk==0.3.13
langgraph==1.1.10
langsmith==0.7.37
librt==0.9.0
mypy==1.20.2
nodeenv==1.10.0
ormsgpack==1.12.2
praw==7.8.1
pre-commit==4.6.0
psycopg[binary]==3.3.3
pylint-django==2.7.0
pylint-plugin-utils==0.9.0
pytest-django==4.12.0
pytest-mock==3.15.1
pytest==9.0.3
python-dotenv==1.2.2
qdrant-client==1.17.1
requests-toolbelt==1.0.0
resend==2.29.0
ruff==0.15.12
sentence-transformers==5.4.1
structlog==25.5.0
types-pyyaml==6.0.12.20260408
uritemplate==4.2.0
uuid-utils==0.14.1
watchdog==6.0.0
xxhash==3.7.0
